Final Submission for Assignment 11.
Three different neural network architectures were evaluated: [10, 32, 1], [10, 64, 32, 1], and [10, 128, 64, 32, 1].
The architecture [10, 32, 1] resulted in an average MSE of approximately 2880.81 on the test set.
The architecture [10, 64, 32, 1] resulted in an average MSE of approximately 2832.07 on the test set.
The architecture [10, 128, 64, 32, 1] resulted in the lowest average MSE of approximately 2653.30 among the tested architectures.
The final model, trained on the entire dataset with the best architecture ([10, 128, 64, 32, 1]), achieved a Mean Squared Error (MSE) of approximately 2163.98 on the test set.
An architecture with more layers and neurons performed best. An even deeper and wider network will potentially yield better results.